+++ cmd/influxd/run/server_helpers_test.go
@@ -78,10 +78,7 @@ func OpenServerWithVersion(c *run.Config, version string) *Server {
 // OpenDefaultServer opens a test server with a default database & retention policy.
 func OpenDefaultServer(c *run.Config) *Server {
 	s := OpenServer(c)
-	if err := s.CreateDatabaseAndRetentionPolicy("db0", newRetentionPolicySpec("rp0", 1, 0)); err != nil {
-		panic(err)
-	}
-	if err := s.MetaClient.SetDefaultRetentionPolicy("db0", "rp0"); err != nil {
+	if err := s.CreateDatabaseAndRetentionPolicy("db0", newRetentionPolicySpec("rp0", 1, 0), true); err != nil {
 		panic(err)
 	}
 	return s
@@ -111,10 +108,10 @@ func (s *Server) URL() string {
 }
 
 // CreateDatabaseAndRetentionPolicy will create the database and retention policy.
-func (s *Server) CreateDatabaseAndRetentionPolicy(db string, rp *meta.RetentionPolicySpec) error {
+func (s *Server) CreateDatabaseAndRetentionPolicy(db string, rp *meta.RetentionPolicySpec, makeDefault bool) error {
 	if _, err := s.MetaClient.CreateDatabase(db); err != nil {
 		return err
-	} else if _, err := s.MetaClient.CreateRetentionPolicy(db, rp); err != nil {
+	} else if _, err := s.MetaClient.CreateRetentionPolicy(db, rp, makeDefault); err != nil {
 		return err
 	}
 	return nil
@@ -455,13 +452,9 @@ func writeTestData(s *Server, t *Test) error {
 			w.rp = t.retentionPolicy()
 		}
 
-		if err := s.CreateDatabaseAndRetentionPolicy(w.db, newRetentionPolicySpec(w.rp, 1, 0)); err != nil {
+		if err := s.CreateDatabaseAndRetentionPolicy(w.db, newRetentionPolicySpec(w.rp, 1, 0), true); err != nil {
 			return err
 		}
-		if err := s.MetaClient.SetDefaultRetentionPolicy(w.db, w.rp); err != nil {
-			return err
-		}
-
 		if res, err := s.Write(w.db, w.rp, w.data, t.params); err != nil {
 			return fmt.Errorf("write #%d: %s", i, err)
 		} else if t.exp != res {