+++ services/meta/client.go
@@ -186,10 +186,7 @@ func (c *Client) CreateDatabase(name string) (*DatabaseInfo, error) {
 	// create default retention policy
 	if c.retentionAutoCreate {
 		rpi := DefaultRetentionPolicyInfo()
-		if err := data.CreateRetentionPolicy(name, rpi); err != nil {
-			return nil, err
-		}
-		if err := data.SetDefaultRetentionPolicy(name, rpi.Name); err != nil {
+		if err := data.CreateRetentionPolicy(name, rpi, true); err != nil {
 			return nil, err
 		}
 	}
@@ -224,7 +221,7 @@ func (c *Client) CreateDatabaseWithRetentionPolicy(name string, spec *RetentionP
 
 	rpi := spec.NewRetentionPolicyInfo()
 	if rp := db.RetentionPolicy(rpi.Name); rp == nil {
-		if err := data.CreateRetentionPolicy(name, rpi); err != nil {
+		if err := data.CreateRetentionPolicy(name, rpi, true); err != nil {
 			return nil, err
 		}
 	} else if !spec.Matches(rp) {
@@ -237,19 +234,17 @@ func (c *Client) CreateDatabaseWithRetentionPolicy(name string, spec *RetentionP
 	// policy we just created. If the default is different from what we are
 	// trying to create, record it as a conflict and abandon with an error.
 	if db.DefaultRetentionPolicy == "" {
-		if err := data.SetDefaultRetentionPolicy(name, rpi.Name); err != nil {
-			return nil, err
-		}
+		db.DefaultRetentionPolicy = rpi.Name
 	} else if rpi.Name != db.DefaultRetentionPolicy {
 		return nil, ErrRetentionPolicyConflict
 	}
 
-	// Refresh the database info.
-	db = data.Database(name)
-
 	if err := c.commit(data); err != nil {
 		return nil, err
 	}
+
+	// Refresh the database info.
+	db = data.Database(name)
 
 	return db, nil
 }
@@ -273,7 +268,7 @@ func (c *Client) DropDatabase(name string) error {
 }
 
 // CreateRetentionPolicy creates a retention policy on the specified database.
-func (c *Client) CreateRetentionPolicy(database string, spec *RetentionPolicySpec) (*RetentionPolicyInfo, error) {
+func (c *Client) CreateRetentionPolicy(database string, spec *RetentionPolicySpec, makeDefault bool) (*RetentionPolicyInfo, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
@@ -284,7 +279,7 @@ func (c *Client) CreateRetentionPolicy(database string, spec *RetentionPolicySpe
 	}
 
 	rp := spec.NewRetentionPolicyInfo()
-	if err := data.CreateRetentionPolicy(database, rp); err != nil {
+	if err := data.CreateRetentionPolicy(database, rp, makeDefault); err != nil {
 		return nil, err
 	}
 
@@ -326,32 +321,14 @@ func (c *Client) DropRetentionPolicy(database, name string) error {
 	return nil
 }
 
-// SetDefaultRetentionPolicy sets a database's default retention policy.
-func (c *Client) SetDefaultRetentionPolicy(database, name string) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	data := c.cacheData.Clone()
-
-	if err := data.SetDefaultRetentionPolicy(database, name); err != nil {
-		return err
-	}
-
-	if err := c.commit(data); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // UpdateRetentionPolicy updates a retention policy.
-func (c *Client) UpdateRetentionPolicy(database, name string, rpu *RetentionPolicyUpdate) error {
+func (c *Client) UpdateRetentionPolicy(database, name string, rpu *RetentionPolicyUpdate, makeDefault bool) error {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
 	data := c.cacheData.Clone()
 
-	if err := data.UpdateRetentionPolicy(database, name, rpu); err != nil {
+	if err := data.UpdateRetentionPolicy(database, name, rpu, makeDefault); err != nil {
 		return err
 	}