+++ services/meta/client_test.go
@@ -248,7 +248,7 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp0.ReplicaN,
 		Duration:           &rp0.Duration,
 		ShardGroupDuration: rp0.ShardGroupDuration,
-	}); err != nil {
+	}, true); err != nil {
 		t.Fatal(err)
 	}
 
@@ -265,7 +265,7 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp0.ReplicaN,
 		Duration:           &rp0.Duration,
 		ShardGroupDuration: rp0.ShardGroupDuration,
-	}); err != nil {
+	}, true); err != nil {
 		t.Fatal(err)
 	} else if actual, err = c.RetentionPolicy("db0", "rp0"); err != nil {
 		t.Fatal(err)
@@ -283,7 +283,7 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp1.ReplicaN,
 		Duration:           &rp1.Duration,
 		ShardGroupDuration: rp1.ShardGroupDuration,
-	})
+	}, true)
 	if exp := meta.ErrRetentionPolicyExists; got != exp {
 		t.Fatalf("got error %v, expected error %v", got, exp)
 	}
@@ -298,7 +298,7 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp1.ReplicaN,
 		Duration:           &rp1.Duration,
 		ShardGroupDuration: rp1.ShardGroupDuration,
-	})
+	}, true)
 	if exp := meta.ErrRetentionPolicyExists; got != exp {
 		t.Fatalf("got error %v, expected error %v", got, exp)
 	}
@@ -313,7 +313,7 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp1.ReplicaN,
 		Duration:           &rp1.Duration,
 		ShardGroupDuration: rp1.ShardGroupDuration,
-	})
+	}, true)
 	if exp := meta.ErrRetentionPolicyExists; got != exp {
 		t.Fatalf("got error %v, expected error %v", got, exp)
 	}
@@ -329,13 +329,13 @@ func TestMetaClient_CreateRetentionPolicy(t *testing.T) {
 		ReplicaN:           &rp1.ReplicaN,
 		Duration:           &rp1.Duration,
 		ShardGroupDuration: rp1.ShardGroupDuration,
-	})
+	}, true)
 	if exp := meta.ErrIncompatibleDurations; got != exp {
 		t.Fatalf("got error %v, expected error %v", got, exp)
 	}
 }
 
-func TestMetaClient_SetDefaultRetentionPolicy(t *testing.T) {
+func TestMetaClient_DefaultRetentionPolicy(t *testing.T) {
 	t.Parallel()
 
 	d, c := newClient()
@@ -402,7 +402,7 @@ func TestMetaClient_UpdateRetentionPolicy(t *testing.T) {
 	if err := c.UpdateRetentionPolicy("db0", "rp0", &meta.RetentionPolicyUpdate{
 		Duration: &duration,
 		ReplicaN: &replicaN,
-	}); err != nil {
+	}, true); err != nil {
 		t.Fatal(err)
 	}
 
@@ -418,7 +418,7 @@ func TestMetaClient_UpdateRetentionPolicy(t *testing.T) {
 	duration = rpi.ShardGroupDuration / 2
 	if err := c.UpdateRetentionPolicy("db0", "rp0", &meta.RetentionPolicyUpdate{
 		Duration: &duration,
-	}); err == nil {
+	}, true); err == nil {
 		t.Fatal("expected error")
 	} else if err != meta.ErrIncompatibleDurations {
 		t.Fatalf("expected error '%s', got '%s'", meta.ErrIncompatibleDurations, err)
@@ -428,7 +428,7 @@ func TestMetaClient_UpdateRetentionPolicy(t *testing.T) {
 	sgDuration := rpi.Duration * 2
 	if err := c.UpdateRetentionPolicy("db0", "rp0", &meta.RetentionPolicyUpdate{
 		ShardGroupDuration: &sgDuration,
-	}); err == nil {
+	}, true); err == nil {
 		t.Fatal("expected error")
 	} else if err != meta.ErrIncompatibleDurations {
 		t.Fatalf("expected error '%s', got '%s'", meta.ErrIncompatibleDurations, err)
@@ -440,7 +440,7 @@ func TestMetaClient_UpdateRetentionPolicy(t *testing.T) {
 	if err := c.UpdateRetentionPolicy("db0", "rp0", &meta.RetentionPolicyUpdate{
 		Duration:           &duration,
 		ShardGroupDuration: &sgDuration,
-	}); err == nil {
+	}, true); err == nil {
 		t.Fatal("expected error")
 	} else if err != meta.ErrIncompatibleDurations {
 		t.Fatalf("expected error '%s', got '%s'", meta.ErrIncompatibleDurations, err)
@@ -452,7 +452,7 @@ func TestMetaClient_UpdateRetentionPolicy(t *testing.T) {
 	if err := c.UpdateRetentionPolicy("db0", "rp0", &meta.RetentionPolicyUpdate{
 		Duration:           &duration,
 		ShardGroupDuration: &sgDuration,
-	}); err != nil {
+	}, true); err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
 }
@@ -481,7 +481,7 @@ func TestMetaClient_DropRetentionPolicy(t *testing.T) {
 		Name:     "rp0",
 		Duration: &duration,
 		ReplicaN: &replicaN,
-	}); err != nil {
+	}, true); err != nil {
 		t.Fatal(err)
 	}